// src/app/interfaces/registration.interface.ts
export interface Registration {
    durationDateFrom: string
    durationDateTo: string
    namePrinted: string
    addressCity: string
    addressStreet: string
    addressState: string
    addressZipcode: string
    phoneNumber: string
    phoneNumber2: string
    vehicle1PlateState: string
    vehicle1PlateNumber: string
    vehicle2PlateState: string
    vehicle2PlateNumber: string
    vehicle3PlateState: string
    vehicle3PlateNumber: string
    vehicle4PlateState: string
    vehicle4PlateNumber: string
    vehicle5PlateState: string
    vehicle5PlateNumber: string
    vehicle6PlateState: string
    vehicle6PlateNumber: string
    vehicle7PlateState: string
    vehicle7PlateNumber: string
    vehicle8PlateState: string
    vehicle8PlateNumber: string
}